*: add util functions to generate random data
