+++ dbms/src/TestUtils/ColumnGenerator.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #include <Columns/ColumnNullable.h>
+#include <Common/RandomData.h>
 #include <DataTypes/DataTypeEnum.h>
 #include <DataTypes/DataTypeNullable.h>
 #include <TestUtils/ColumnGenerator.h>
@@ -31,7 +32,6 @@ ColumnWithTypeAndName ColumnGenerator::generateNullMapColumn(const ColumnGenerat
 
 ColumnWithTypeAndName ColumnGenerator::generate(const ColumnGeneratorOpts & opts)
 {
-    int_rand_gen = std::uniform_int_distribution<Int64>(0, opts.string_max_size);
     DataTypePtr type;
     if (opts.type_name == "Decimal")
         type = createDecimalType();
@@ -98,7 +98,7 @@ ColumnWithTypeAndName ColumnGenerator::generate(const ColumnGeneratorOpts & opts
         break;
     case TypeIndex::String:
         for (size_t i = 0; i < opts.size; ++i)
-            genString(col);
+            genString(col, opts.string_max_size);
         break;
     case TypeIndex::Decimal32:
     case TypeIndex::Decimal64:
@@ -124,8 +124,6 @@ ColumnWithTypeAndName ColumnGenerator::generate(const ColumnGeneratorOpts & opts
         for (size_t i = 0; i < opts.size; ++i)
             genEnumValue(col, type);
         break;
-        {
-        }
     default:
         throw std::invalid_argument("RandomColumnGenerator invalid type");
     }
@@ -139,73 +137,6 @@ DataTypePtr ColumnGenerator::createDecimalType()
     int prec = rand_gen() % max_precision + 1;
     int scale = rand_gen() % prec;
     return DB::createDecimal(prec, scale);
-}
-
-String ColumnGenerator::randomString()
-{
-    String str(int_rand_gen(rand_gen), 0);
-    std::generate_n(str.begin(), str.size(), [this]() { return charset[rand_gen() % charset.size()]; });
-    return str;
-}
-
-int ColumnGenerator::randomTimeOffset()
-{
-    static constexpr int max_offset = 24 * 3600 * 10000; // 10000 days for test
-    return (rand_gen() % max_offset) * (rand_gen() % 2 == 0 ? 1 : -1);
-}
-
-time_t ColumnGenerator::randomUTCTimestamp()
-{
-    return ::time(nullptr) + randomTimeOffset();
-}
-
-struct tm ColumnGenerator::randomLocalTime()
-{
-    time_t t = randomUTCTimestamp();
-    struct tm res
-    {
-    };
-
-    if (localtime_r(&t, &res) == nullptr)
-    {
-        throw std::invalid_argument(fmt::format("localtime_r({}) ret {}", t, strerror(errno)));
-    }
-    return res;
-}
-
-String ColumnGenerator::randomDate()
-{
-    auto res = randomLocalTime();
-    return fmt::format("{}-{}-{}", res.tm_year + 1900, res.tm_mon + 1, res.tm_mday);
-}
-
-String ColumnGenerator::randomDuration()
-{
-    auto res = randomLocalTime();
-    return fmt::format("{}:{}:{}", res.tm_hour, res.tm_min, res.tm_sec);
-}
-
-String ColumnGenerator::randomDateTime()
-{
-    auto res = randomLocalTime();
-    return fmt::format(
-        "{}-{}-{} {}:{}:{}",
-        res.tm_year + 1900,
-        res.tm_mon + 1,
-        res.tm_mday,
-        res.tm_hour,
-        res.tm_min,
-        res.tm_sec);
-}
-
-String ColumnGenerator::randomDecimal(uint64_t prec, uint64_t scale)
-{
-    auto s = std::to_string(rand_gen());
-    if (s.size() < prec)
-        s += String(prec - s.size(), '0');
-    else if (s.size() > prec)
-        s = s.substr(0, prec);
-    return s.substr(0, prec - scale) + "." + s.substr(prec - scale);
 }
 
 template <typename IntegerType>
@@ -264,9 +195,9 @@ void ColumnGenerator::genFloat(MutableColumnPtr & col)
     col->insert(f);
 }
 
-void ColumnGenerator::genString(MutableColumnPtr & col)
+void ColumnGenerator::genString(MutableColumnPtr & col, UInt64 max_size)
 {
-    Field f = randomString();
+    Field f = randomString(max_size);
     col->insert(f);
 }
